search: extract DynamicFilter code from SearchResultsResolver
This will allow us to compute search filters during a stream. Currently
we have to wait until all results are found. Additionally it removes
another dependency on aggregation to do streaming.
